chore: remove unwanted code
